requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
arxiv==2.1.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.130
botocore==1.34.130
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.8
colorama==0.4.6
cryptography==43.0.0
cssselect==1.2.0
cssselect2==0.7.0
dataclasses-json==0.6.7
deepdiff==7.0.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
duckduckgo-search @ https://files.pythonhosted.org/packages/e1/c5/5eaff940cf0d9a6d080397d07d7071eafa6c191af5e34f1da9776b3094f9/duckduckgo_search-6.2.12-py3-none-any.whl
emoji==2.12.1
fastavro==1.9.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.15.3
filetype==1.2.0
firebase-admin==6.5.0
fireworks-ai==0.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.56.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-aws==0.1.6
langchain-cohere==0.1.8
langchain-community==0.2.5
langchain-core==0.2.9
langchain-fireworks==0.1.3
langchain-google-genai==1.0.6
langchain-google-vertexai==1.0.5
langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-mistralai==0.1.8
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchain-together==0.1.3
langdetect==1.0.9
langsmith==0.1.81
lxml==5.2.2
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.3
md2pdf==1.0.1
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.35.1
ordered-set==4.1.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.3.0
playwright==1.44.0
primp==0.6.2
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydyf==0.10.0
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
pyphen==0.15.0
pyreqwest_impersonate==0.4.8
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sgmllib3k==1.0.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
streamlit==1.36.0
sympy==1.12.1
tabulate==0.9.0
tavily-python==0.3.3
tbb==2021.13.0
tenacity==8.4.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.2
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.14.6
unstructured-client==0.23.7
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
weasyprint==62.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zopfli==0.2.3
